move db_version test to storage/db tests
